requirements.txt
redis
rq
djangorestframework
flask==1.1.1
paho-mqtt==1.5.0
sqlalchemy==1.3.13
pymysql==0.9.3